+++ src/test/groovy/com/cedarsoftware/ncube/TestNCubeManager.groovy
@@ -600,7 +600,7 @@ class TestNCubeManager extends NCubeCleanupBaseTest
             assertContainsIgnoreCase(e.message, '1.0.0 already exists')
         }
 
-        assert 2 == mutableClient.releaseCubes(defaultSnapshotApp, '1.2.3')
+        assert 3 == mutableClient.releaseCubes(defaultSnapshotApp, '1.2.3') // 2 + sys.info = 3
 
         try
         {
@@ -1772,6 +1772,73 @@ class TestNCubeManager extends NCubeCleanupBaseTest
         assertEquals(2, userColumns.size())
         assertEquals(System.getProperty('user.name'), userColumns.get(0).value)
         assertEquals(3, sysUsergroupsCube.getAxis(AXIS_ROLE).columns.size())
+    }
+
+    @Test
+    void testCreateCubeCreatesSysInfoCubes()
+    {
+        NCube sysInfo = mutableClient.getCube(ApplicationID.testAppId, SYS_INFO)
+        assert 1 == sysInfo.numDimensions
+        Axis attribute = sysInfo.getAxis(AXIS_ATTRIBUTE)
+        assert 1 == attribute.size()
+        assert attribute.hasDefaultColumn()
+
+        NCube sysInfo000 = mutableClient.getCube(ApplicationID.testAppId.asVersion(SYS_BOOT_VERSION), SYS_INFO)
+        assert 1 == sysInfo000.numDimensions
+        Axis attribute000 = sysInfo000.getAxis(AXIS_ATTRIBUTE)
+        assert 1 == attribute000.size()
+        assert attribute000.hasDefaultColumn()
+    }
+
+    @Test
+    void testUnableToDeleteSysInfo()
+    {
+        NCube sysClasspath = mutableClient.getCube(ApplicationID.testAppId, SYS_CLASSPATH)
+        NCube sysInfo = mutableClient.getCube(ApplicationID.testAppId, SYS_INFO)
+        assert sysClasspath != null
+        assert sysInfo != null
+
+        mutableClient.deleteCubes(ApplicationID.testAppId, [SYS_INFO, SYS_CLASSPATH] as String[])
+
+        sysClasspath = mutableClient.getCube(ApplicationID.testAppId, SYS_CLASSPATH)
+        sysInfo = mutableClient.getCube(ApplicationID.testAppId, SYS_INFO)
+        assert sysClasspath == null
+        assert sysInfo != null
+    }
+
+    @Test
+    void testDuplicateCreatesSysInfo()
+    {
+        ApplicationID newApp = ApplicationID.testAppId.asBranch('TEST2')
+        mutableClient.duplicate(ApplicationID.testAppId, newApp, SYS_CLASSPATH, SYS_CLASSPATH)
+        NCube sysClasspath = mutableClient.getCube(newApp, SYS_CLASSPATH)
+        NCube sysInfo = mutableClient.getCube(newApp, SYS_INFO)
+        assert sysClasspath != null
+        assert sysInfo != null
+    }
+
+    @Test
+    void testCommitCubesCreatesSysInfo()
+    {
+        ApplicationID headApp = ApplicationID.testAppId.asHead()
+        NCube headSysInfo = mutableClient.getCube(headApp, SYS_INFO)
+        NCube headSysInfo000 = mutableClient.getCube(headApp.asVersion(SYS_BOOT_VERSION), SYS_INFO)
+        assert headSysInfo == null
+        assert headSysInfo000 != null // sys.info is created in 0.0.0 HEAD when permissions n-cubes are created
+
+        mutableClient.commitBranch(ApplicationID.testAppId)
+        headSysInfo = mutableClient.getCube(headApp, SYS_INFO)
+        headSysInfo000 = mutableClient.getCube(headApp.asVersion(SYS_BOOT_VERSION), SYS_INFO)
+        assert headSysInfo != null
+        assert headSysInfo000 != null
+    }
+
+    @Test
+    void testSearchNotIncludeSysInfo()
+    {
+        List<NCubeInfoDto> infoDtos = mutableClient.search(ApplicationID.testAppId, null, null, null)
+        assert 1 == infoDtos.size()
+        assert SYS_CLASSPATH == infoDtos[0].name
     }
 
     @Test